Split PlaygroundAi component into smaller chunks
